requirements.txt
elasticsearch>=8.12.0,<9.0.0
aiohttp>=3.9.0
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.27.0
groq>=0.4.0
python-dotenv>=1.0.0
rich>=13.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
